During powersave (on) first call, still set brightness
If we shutdown the laptop in unpowered state, systemd-backlight will
restore the dimmed brightness setting at boot, even if we are powered on.
So during the first call to powersave, where normally we do nothing except
storing the powersave settings, we still set the brightness.

We could also pass systemd.restore_state=0 in the kernel to prevent
systemd-backlight to restore the last brightness, but this would also
prevent systemd-rfkill (which restore rfkill states) from being called.

Maybe it would be easier to just go ahead with the whole powersave program,
even during the first call.